# rest/src/main/resources/openapi/admin/v7.0/openapi.yaml
openapi: 3.0.2
info:
  title: Administrační API v systému Kramerius
  description: |
    Popis je platný k verzi 7.0.38 a vyšší. Kramerius obsahuje REST API, které umožňuje všechny operace nad systémem.   Standardně je dostupné na nainstalovaném serveru pod kontextem: `~/search/api/admin/v7.0*`.  
    Primárně určeno pro administraci  

  version: "7.0"
servers:
  - url: /search/api/admin/v7.0
...
paths:
...
  /licenses:
    get:
      tags:
        - Licence
      summary: Vrací všechny licence
      operationId: getAllLicenses
      description: Vrací všechny licence
      parameters:
        - name: runtime
          in: query
          required: true
          schema:
            type: boolean
            default: true
          description: Zda je nutné zahrnout do výpisu i runtime licence

      responses:
        '200':
          description: JSON reprezentace licencí
          content:
            application/json:
              schema:
                type: string
              example: [{
                        "name": "dnntt",
                        "description": "DNNT terminal license",
                        "id": 1,
                        "priority": 1,
                        "group": "embedded"
                    },
                    {
                        "name": "dnnto",
                        "description": "DNNT online license",
                        "id": 2,
                        "priority": 2,
                        "group": "embedded"
                    }]
...
  /repository/getMetadata:
    get:
      summary: Získání metadat objecktu z repozitáře.
      description: Získá metadata objektu z akubra repozitáře .
      operationId: getMetadata
      parameters:
        - name: pid
          in: query
          required: true
          description: Identifikátor objektu.
          schema:
            type: string
      responses:
        '200':
          description: Úspěšné získání metadat objektu
          content:
            application/json:
              schema:
                type: object
                properties:
                  datastream:
                    type: object
                    properties:
                      propertyLabel:
                        type: string
                        description: label property
                      propertyCreated:
                        type: string
                        format: date-time
                        description: Datum vytvoření objektu
                      propertyLastModified:
                        type: string
                        format: date-time
                        description: Datum oprava objektu
                      objectStoragePath:
                        type: string
                        description: Object storage path
        '400':
          description: Nevalidní parametry http dotazu
        '404':
          description: Datastream nebo objekt není nalezen
      tags:
        - akubra
...
  /ui-config/metadata:
    get:
      tags:
        - UIConfig
      summary: Vraci metadata editoru UI konfigurace
      description: >
        Vraci seznam UI konfiguraci vcetne endpointu, titulku, schematu a patternu
        pro detekci resource referenci. Patterny jsou sestavene z defaultu a KConfiguration.
      responses:
        '200':
          description: Metadata UI konfigurace
          content:
            application/json:
              schema:
                type: object
                properties:
                  configs:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                          example: ui-config.general
                        endpoint:
                          type: string
                          example: general
                        title:
                          type: string
                          example: General JSON
                        schema:
                          type: object
                          properties:
                            type:
                              type: string
                              example: object
                        resourceReferencePatterns:
                          type: array
                          items:
                            type: string
                          example:
                            - /pages/content/*
        '403':
          description: Uzivatel nema opravneni k pozadovane akci
  /ui-config/general:
    get:
      tags:
        - UIConfig
      summary: Vrací JSON vseobecnou konfiguraci klienta
      description: Vrací JSON vseobecnou konfiguraci klienta
      responses:
        '200':
          description: JSON reprezentace vseobecne konfigurace
          content:
            application/json:
              schema:
                type: string
                example: {
                    "share_url": "https://kramerius.lib.cas.cz/uuid/${UUID}",
                    "ga4": "G-3DSV45X30Q",
                    "ga4clientId": "kramerius.lib.cas.cz",
                    "enablePeriodicalVolumesYearsLayout": false,
                    "enablePeriodicalIsssuesCalendarLayout": false,
                    "defaultPeriodicalVolumesLayout": "grid",
                    "defaultPeriodicalIssuesLayout": "grid",
                    "publicFilterDefault": false,
                    "bigHomeLogo": false,
                    "hideHomeTitle": false
                }
        '403':
          description: Uživatel nemá oprávnění k požadované akci
        '404':
          description: UI config neexistuje
    post:
      tags:
        - UIConfig
      summary: Vložení JSON nové vseobecne konfigurace
      description: Vložení JSON nové vseobecne konfigurace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
            example: {
              "share_url": "https://kramerius.lib.cas.cz/uuid/${UUID}",
              "ga4": "G-3DSV45X30Q",
              "ga4clientId": "kramerius.lib.cas.cz",
              "enablePeriodicalVolumesYearsLayout": false,
              "enablePeriodicalIsssuesCalendarLayout": false,
              "defaultPeriodicalVolumesLayout": "grid",
              "defaultPeriodicalIssuesLayout": "grid",
              "publicFilterDefault": false,
              "bigHomeLogo": false,
              "hideHomeTitle": false
            }
      responses:
        '201':
          description: Konfigurace úspěšně vytvořena
        '400':
          description: Neplatná data požadavku
        '403':
          description: Uživatel nemá oprávnění k požadované akci
  /ui-config/licenses:
    get:
      tags:
        - UIConfig
      summary: Vrací JSON konfiguraci klienta tykajici se licenci
      description: Vrací JSON konfiguraci klienta tykajici se licenci
      responses:
        '200':
          description: JSON reprezentace konfigurace licenci
          content:
            application/json:
              schema:
                type: string
                example: {
                  "public": {
                    "access": "open",
                    "label": {
                      "cs": "Volna dila",
                      "en": "Public domain"
                    },
                    "message": {
                      "cs": "/assets/shared/licences/public.cs.html",
                      "en": "/assets/shared/licences/public.en.html"
                    },
                    "instruction": {
                      "cs": "/assets/shared/licences/public.instruction.cs.html",
                      "en": "/assets/shared/licences/public.instruction.en.html"
                    }
                  },
                  "orphan": {
                    "access": "open",
                    "label": {
                      "cs": "Osirela dila",
                      "en": "Orphan works"
                    },
                    "message": {
                      "cs": "/assets/shared/licences/orphan.cs.html",
                      "en": "/assets/shared/licences/orphan.en.html"
                    },
                    "instruction": {
                      "cs": "/assets/shared/licences/orphan.instruction.cs.html",
                      "en": "/assets/shared/licences/orphan.instruction.en.html"
                    }
                  }
                }
        '403':
          description: Uživatel nemá oprávnění k požadované akci
        '404':
          description: UI config neexistuje
    post:
      tags:
        - UIConfig
      summary: Vložení JSON nové konfigurace tykajici se licenci
      description: Vložení JSON nové konfigurace tykajici se licenci
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
            example: {
              "public": {
                "access": "open",
                "label": {
                  "cs": "Volna dila",
                  "en": "Public domain"
                },
                "message": {
                  "cs": "/assets/shared/licences/public.cs.html",
                  "en": "/assets/shared/licences/public.en.html"
                },
                "instruction": {
                  "cs": "/assets/shared/licences/public.instruction.cs.html",
                  "en": "/assets/shared/licences/public.instruction.en.html"
                }
              },
              "orphan": {
                "access": "open",
                "label": {
                  "cs": "Osirela dila",
                  "en": "Orphan works"
                },
                "message": {
                  "cs": "/assets/shared/licences/orphan.cs.html",
                  "en": "/assets/shared/licences/orphan.en.html"
                },
                "instruction": {
                  "cs": "/assets/shared/licences/orphan.instruction.cs.html",
                  "en": "/assets/shared/licences/orphan.instruction.en.html"
                }
              }
            }
      responses:
        '201':
          description: Konfigurace úspěšně vytvořena
        '400':
          description: Neplatná data požadavku
        '403':
          description: Uživatel nemá oprávnění k požadované akci
  /ui-config/curator-lists:
    get:
      tags:
        - UIConfig
      summary: Vrací JSON konfiguraci klienta tykajici se kuratorskych seznamu
      description: Vrací JSON konfiguraci klienta tykajici se kuratorskych seznamu
      responses:
        '200':
          description: JSON reprezentace konfigurace licenci
          content:
            application/json:
              schema:
                type: string
                example: [
                  {
                    "cs": "Casopisy Akademie ved CSR",
                    "en": "Journals of the Czech Academy of Sciences",
                    "content": [
                      "uuid:694b7069-9b46-46a9-8665-2b1ea226172c",
                      "uuid:b7c97808-0502-4246-bd07-88108d461626",
                      "uuid:c82a1017-48e1-11e1-9389-005056a60003"
                    ]
                  },
                  {
                    "cs": "Vybrane sbirky",
                    "en": "Selected collections",
                    "content": [
                      "uuid:30d8c0e8-4e26-4453-83c7-d38a4e178805",
                      "uuid:175f7bc8-e51b-4fb3-b737-cfd7f3f2a43b",
                      "uuid:314bf0c1-9863-442e-8439-b3fe1cec88ae"
                    ]
                  }
                ]
        '403':
          description: Uživatel nemá oprávnění k požadované akci
        '404':
          description: UI config neexistuje
    post:
      tags:
        - UIConfig
      summary: Vložení JSON nové konfigurace tykajici se kuratorskych seznamu
      description: Vložení JSON nové konfigurace tykajici se kuratorskych seznamu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
            example: [
              {
                "cs": "Casopisy Akademie ved CSR",
                "en": "Journals of the Czech Academy of Sciences",
                "content": [
                  "uuid:694b7069-9b46-46a9-8665-2b1ea226172c",
                  "uuid:b7c97808-0502-4246-bd07-88108d461626",
                  "uuid:c82a1017-48e1-11e1-9389-005056a60003"
                ]
              },
              {
                "cs": "Vybrane sbirky",
                "en": "Selected collections",
                "content": [
                  "uuid:30d8c0e8-4e26-4453-83c7-d38a4e178805",
                  "uuid:175f7bc8-e51b-4fb3-b737-cfd7f3f2a43b",
                  "uuid:314bf0c1-9863-442e-8439-b3fe1cec88ae"
                ]
              }
            ]
      responses:
        '201':
          description: Konfigurace úspěšně vytvořena
        '400':
          description: Neplatná data požadavku
        '403':
          description: Uživatel nemá oprávnění k požadované akci
...